added hostel pagination and hostel rating update method

+++ src/main/java/org/example/hostelbooking/web/entity/hostel/HostelResponse.java
@@ -25,7 +25,7 @@ public class HostelResponse {
 
     private Float distance;
 
-    private Integer rating;
+    private Float rating;
 
     private Float ratingCounter;
